Cleanups.

+++ src/context/PidControlProvider.tsx
@@ -16,11 +16,10 @@ type Props = {
   children: React.ReactNode;
 };
 
-const KP = parseInt(import.meta.env.VITE_PID_KP || '1');
-const KI = parseInt(import.meta.env.VITE_PID_KI || '0');
-const KD = parseInt(import.meta.env.VITE_PID_KD || '0.1');
-// Dummy:  kp: 20, ki: 0.6, kd: 1,
-// OwnMachine:  kp: 1.5, ki: 0.012, kd: 12
+const KP = parseFloat(import.meta.env.VITE_PID_KP || '1');
+const KI = parseFloat(import.meta.env.VITE_PID_KI || '0');
+const KD = parseFloat(import.meta.env.VITE_PID_KD || '0.1');
+
 export const PidControlProvider: React.FC<Props> = ({ children }) => {
   const [setpoint, setSetpoint] = useState<number>(0);
   const [enabled, setEnabled] = useState(true);
